内容の把握だけ
gamemanagerのcheckgameoverでゲームオーバーを管理してると推測されます。
他ファイルから値の参照がうまくいけば、ゲームオーバーの遷移もうまくいきそうです。
