--- pinball_assets_ver/Assets/Flipper/Lflipper.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: eb43c35a484c64381a6d2e1e96afc51f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
