+++ pinball_assets_ver/Assets/_Pinball/Scenes/Main.unity
@@ -8727,18 +8727,6 @@ GameObject:
     type: 3}
   m_PrefabInstance: {fileID: 2750021993519768125}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &2052558111
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2052558106}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: eb43c35a484c64381a6d2e1e96afc51f, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &2057103269
 GameObject:
   m_ObjectHideFlags: 0
@@ -9079,6 +9067,8 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   slider: {fileID: 2142147718}
+  gameManager: {fileID: 0}
+  ballController: {fileID: 0}
 --- !u!212 &2127545579
 SpriteRenderer:
   m_ObjectHideFlags: 0
