--- pinball_assets_ver/Assets/image/PlayerController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 355f86739333c436d9ca81bb1b881c0e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
