--- pinball_assets_ver/Library/ScriptAssemblies/BuiltinAssemblies.stamp
@@ -1,1 +1,0 @@
-08d9d0a4af115b00.08d9d0a4c3555800
