+++ pinball_assets_ver/Library/StateCache/SceneView/8c/8cd7c613bf844de3b80696e27a479d5e.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"m_AlwaysRefresh":false,"showFog":true,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":true,"pivot":{"x":-0.004190757405012846,"y":-0.018758626654744149,"z":-10.000423431396485},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":12.151628494262696,"orthographic":true}
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"m_AlwaysRefresh":false,"showFog":true,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":true,"pivot":{"x":161.988525390625,"y":52.94561004638672,"z":-520.2265625},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":82.93082427978516,"orthographic":true}
