+++ pinball_assets_ver/Logs/shadercompiler-UnityShaderCompiler0.log
@@ -1,3 +1,3 @@
-Base path: '/Applications/Unity/Unity.app/Contents', plugins path '/Applications/Unity/Unity.app/Contents/PlaybackEngines'
+Base path: '/Users/kodamahideaki/Desktop/2020.3.26f1/Unity.app/Contents', plugins path '/Users/kodamahideaki/Desktop/2020.3.26f1/Unity.app/Contents/PlaybackEngines'
 Cmd: initializeCompiler
